+++ lib-database-entities/src/commonMain/kotlin/com/ustadmobile/lib/db/entities/DownloadJobItemAndNodeInfo.kt
@@ -2,7 +2,7 @@ package com.ustadmobile.lib.db.entities
 
 class DownloadJobItemAndNodeInfo : DownloadJobItem(){
 
-    var networkNodeId: Long = 0
+    var networkNodeId: Int = 0
 
     var networkNodeEndpointUrl: String? = null
