+++ retriever-runtime/src/commonMain/kotlin/com/ustadmobile/retriever/AvailabilityEvent.kt
@@ -7,7 +7,7 @@ class AvailabilityEvent(
      */
     val availabilityInfo: Map<String, AvailabilityEventInfo>,
 
-    val networkNodeUid: Long,
+    val networkNodeUid: Int,
 
     /**
      * If true, then there are still more local nodes that we don't have an up-to-date response from. If false, we have
