+++ retriever-runtime/src/commonMain/kotlin/com/ustadmobile/retriever/RetrieverStatusUpdateEvent.kt
@@ -1,7 +1,7 @@
 package com.ustadmobile.retriever
 
 data class RetrieverStatusUpdateEvent(
-    val downloadJobItemUid: Long,
+    val downloadJobItemUid: Int,
     val url: String,
     val status: Int,
 ) {
